LayoutTests/http/tests/security/contentSecurityPolicy/allow-inline-remove-attribute-expected.txt

CONSOLE MESSAGE: Refused to apply a stylesheet because its hash, its nonce, or 'unsafe-inline' does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: PASS: style attribute successfully removed.
